backend/requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anthropic==0.68.0
anyio==4.10.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
docstring_parser==0.17.0
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.60.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.6
numpy==2.3.3
openai==1.108.1
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
stripe==12.5.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
openai==0.28.1

# Oracle
oracledb==2.4.1

openpyxl>=3.1.5
skorecard==1.4.0
category-encoders>=2.6,<3.0
